- removed todo's
git-svn-id: svn://141.76.65.164/kp-salespoint@367 eb53f5e8-9ec5-4bc0-8d13-06c033283886

+++ SalespointFramework/src/org/salespointframework/core/order/PersistentOrderManager.java
@@ -140,19 +140,19 @@ public class PersistentOrderManager implements OrderManager {
 	 * @see org.salespointframework.core.order.IOrderManager#removeOrder(org.
 	 * salespointframework.core.order.OrderIdentifier)
 	 */
-	@Override
-	public OrderEntry removeOrder(OrderIdentifier orderIdentifier) {
-		Objects.requireNonNull(orderIdentifier, "orderIdentifier");
+    @Override
+    public OrderEntry removeOrder(OrderIdentifier orderIdentifier) {
+        Objects.requireNonNull(orderIdentifier, "orderIdentifier");
 
-		em.getTransaction().begin();
-		
-		OrderEntry orderEntry = em.find(OrderEntry.class, orderIdentifier);
-		//TODO: what if orderEntry == null?!
-		em.remove(orderEntry);
-		
-		em.getTransaction().commit();
-		return orderEntry;
-	}
+        OrderEntry orderEntry = em.find(OrderEntry.class, orderIdentifier);
+
+        if (orderEntry != null) {
+            em.getTransaction().begin();
+            em.remove(orderEntry);
+            em.getTransaction().commit();
+        }
+        return orderEntry;
+    }
 
 	/*
 	 * (non-Javadoc)
